pyproject.toml
[project]
name = "llm-knowledge-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-tavily>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.38.0",
    "watchdog>=4.0.0",
    "wikipedia>=1.4.0",
]

[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
select = ["F", "E", "W", "I"]


[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.10",
]
